+++ src/apps_control.rs
@@ -36,10 +36,12 @@ pub fn start_app(id: AppId) -> Result<String, String> {
             Ok("Started RabbitMQ.".into())
         }
         AppId::Phpmyadmin => Err("phpMyAdmin does not support Start/Stop from Apps.".into()),
-        AppId::Snappymail | AppId::Tachyon | AppId::Nextsnapmail | AppId::Sogo => Err(format!(
-            "{} does not support Start/Stop from Apps.",
-            id.label()
-        )),
+        AppId::Snappymail | AppId::Tachyon | AppId::Nextcloud | AppId::Nextsnapmail | AppId::Sogo => {
+            Err(format!(
+                "{} does not support Start/Stop from Apps.",
+                id.label()
+            ))
+        }
     }
 }
 
@@ -69,9 +71,11 @@ pub fn stop_app(id: AppId) -> Result<String, String> {
             Ok("Stopped RabbitMQ.".into())
         }
         AppId::Phpmyadmin => Err("phpMyAdmin does not support Start/Stop from Apps.".into()),
-        AppId::Snappymail | AppId::Tachyon | AppId::Nextsnapmail | AppId::Sogo => Err(format!(
-            "{} does not support Start/Stop from Apps.",
-            id.label()
-        )),
+        AppId::Snappymail | AppId::Tachyon | AppId::Nextcloud | AppId::Nextsnapmail | AppId::Sogo => {
+            Err(format!(
+                "{} does not support Start/Stop from Apps.",
+                id.label()
+            ))
+        }
     }
 }